Create routes and curl scripts
This commit creates the router for ratings.  It also contains curl
scripts for testing.  The router allows the user to make requests to
create, show, update, or destroy their entries into the database. The
only user who can make changes on their entries is the actual owner of
that file.