Dev (#2)
* Create the model

This commit contains the completed rating model for version 1 of the
ratings app.  It contains all of the data fields, the owner, and the
timestamps virtual

* New commit for rating model

The previous commit did not contain the intended resources, this is The
commit that does contain it

* Create routes and curl scripts

This commit creates the router for ratings.  It also contains curl
scripts for testing.  The router allows the user to make requests to
create, show, update, or destroy their entries into the database. The
only user who can make changes on their entries is the actual owner of
that file.

* Modify the curl script for index

This includes a modifcation to the index.sh file

* Changes the status and data

This sends back a status code and body when updated